requirements.txt
Django==2.2.1
django-filter==2.1.0
django-tables2==2.0.6
pytz==2019.1
sqlparse==0.3.0
